+++ app/styles/container.module.scss
@@ -1,4 +1,4 @@
-@import "colors";
+@import "./utils/colors";
 
 .main-container {
   background-color: $red;
