+++ app/styles/globals.scss
@@ -1,4 +1,4 @@
-@import "colors";
+@import "./utils/colors";
 
 html,
 body {
