+++ app/styles/layout.module.scss
@@ -1,4 +1,4 @@
-@import "mixins";
+@import "./utils/mixins";
 
 .container {
   display: flex;
