+++ app/styles/nav.module.scss
@@ -1,5 +1,5 @@
-@import "colors";
-@import "mixins";
+@import "./utils/colors";
+@import "./utils/mixins";
 
 .main-nav {
   background-color: $grey;
