--- barba-con-bigote-redesign/styles/utils/_colors.scss
@@ -1,32 +1,0 @@
-// Brand Colors
-$brown: #3F2016;
-$grey: #EFEFEF;
-$orange: #F79425;
-$red: #D54935;
-$white: #ffffff;
-
-// Behance
-$behance-white: #F9F9F9;
-$behance-black: #191919;
-$behance-blue: #0057FF;
-
-// Dribbble
-$dribbble-pink: #EA4C89;
-$dribbble-black: #0D0C22;
-$dribbble-blue: #00B6E3;
-
-// Github
-$github-white: #F4F7FB;
-$github-black: #0E0D0D;
-$github-purple: #6F42C1;
-
-// Cerveza
-$cerveza-crema: #F8EFE7;
-$ceveza-negro: #2C2D48;
-
-// Spotify
-$spotify-black: #191414;
-$spotify-green: #22D760;
-
-// Shadows
-$tooltip-shadow: 0 2px 4px rgba(63, 32, 22, .2);
