--- barba-con-bigote-redesign/styles/utils/_mixins.scss
@@ -1,35 +1,0 @@
-// Variables
-$sm: 448px;
-$md: 784px;
-$lg: 1024px;
-$xl: 1280px;
-$xxl: 1536px;
-
-// Media query
-@mixin media($size) {
-  @if $size == sm {
-    @media (min-width: $sm) { 
-      @content; 
-    }
-  } 
-  @else if $size == md {
-    @media (min-width: $md) {
-      @content;
-    }
-  } 
-  @else if $size == lg {
-    @media (min-width: $lg) {
-      @content;
-    }  
-  } 
-  @else if $size == xl {
-    @media (min-width: $xl) {
-      @content;
-    } 
-  }
-  @else if $size == xxl {
-    @media (min-width: $xxl) {
-      @content;
-    }
-  }
-}
